added reusable material ui custom table class
