+++ package.json
@@ -16,6 +16,7 @@
   },
   "dependencies": {
     "@material-ui/core": "^4.9.1",
+    "@material-ui/icons": "^4.9.1",
     "@types/loader-utils": "^1.1.3",
     "apisauce": "^1.1.1",
     "body-parser": "^1.17.2",
