+++ src/components/Utility/CustomTable/index.js
@@ -1,0 +1,1 @@
+export { default } from './CustomTable';
